Fixes CI
